pyproject.toml
[tool.poetry]
name = "bewise-test-case"
version = "0.1.0"
description = ""
authors = ["ikhanter <khantermail@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
